moved network logic inside Announcement, moved Announcement component
